refactor(server): extract reusable run_server library entrypoint
